chore(source): code cleanup to reduce coupling
