+++ source/istio_gateway.go
@@ -19,6 +19,7 @@ package source
 import (
 	"context"
 	"fmt"
+	"reflect"
 	"sort"
 	"strings"
 	"text/template"
@@ -28,6 +29,7 @@ import (
 	istioclient "istio.io/client-go/pkg/clientset/versioned"
 	istioinformers "istio.io/client-go/pkg/informers/externalversions"
 	networkingv1alpha3informer "istio.io/client-go/pkg/informers/externalversions/networking/v1alpha3"
+	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	kubeinformers "k8s.io/client-go/informers"
@@ -36,11 +38,15 @@ import (
 	"k8s.io/client-go/tools/cache"
 
 	"sigs.k8s.io/external-dns/endpoint"
+	"sigs.k8s.io/external-dns/source/utils"
 )
 
 // IstioGatewayIngressSource is the annotation used to determine if the gateway is implemented by an Ingress object
 // instead of a standard LoadBalancer service type
-const IstioGatewayIngressSource = "external-dns.alpha.kubernetes.io/ingress"
+const (
+	IstioGatewayIngressSource = "external-dns.alpha.kubernetes.io/ingress"
+	ApiGatewayV1Alpha3        = "gateway.networking.istio.io/v1alpha3"
+)
 
 // gatewaySource is an implementation of Source for Istio Gateway objects.
 // The gateway implementation uses the spec.servers.hosts values for the hostnames.
@@ -84,7 +90,16 @@ func NewIstioGatewaySource(
 	serviceInformer.Informer().AddEventHandler(
 		cache.ResourceEventHandlerFuncs{
 			AddFunc: func(obj interface{}) {
-				log.Debug("service added")
+				// TODO: test this
+				if log.IsLevelEnabled(log.DebugLevel) {
+					service, ok := obj.(*corev1.Service)
+					if !ok {
+						log.Errorf("event handler not added. want 'service/v1' got '%s'", reflect.TypeOf(obj).String())
+						return
+					} else {
+						log.Debugf("event handler added for 'service/v1' in 'namespace:%s' with 'name:%s'.", service.Name, service.Namespace)
+					}
+				}
 			},
 		},
 	)
@@ -92,7 +107,16 @@ func NewIstioGatewaySource(
 	gatewayInformer.Informer().AddEventHandler(
 		cache.ResourceEventHandlerFuncs{
 			AddFunc: func(obj interface{}) {
-				log.Debug("gateway added")
+				// TODO: test this
+				if log.IsLevelEnabled(log.DebugLevel) {
+					service, ok := obj.(*networkingv1alpha3.Gateway)
+					if !ok {
+						log.Errorf("event handler not added. want '%s' got '%s'", ApiGatewayV1Alpha3, reflect.TypeOf(obj).String())
+						return
+					} else {
+						log.Debugf("event handler added for '%s' in 'namespace:%s' with 'name:%s'", ApiGatewayV1Alpha3, service.Name, service.Namespace)
+					}
+				}
 			},
 		},
 	)
@@ -200,11 +224,7 @@ func (sc *gatewaySource) AddEventHandler(ctx context.Context, handler func()) {
 
 // filterByAnnotations filters a list of configs by a given annotation selector.
 func (sc *gatewaySource) filterByAnnotations(gateways []*networkingv1alpha3.Gateway) ([]*networkingv1alpha3.Gateway, error) {
-	labelSelector, err := metav1.ParseToLabelSelector(sc.annotationFilter)
-	if err != nil {
-		return nil, err
-	}
-	selector, err := metav1.LabelSelectorAsSelector(labelSelector)
+	selector, err := utils.ParseAnnotationFilter(sc.annotationFilter)
 	if err != nil {
 		return nil, err
 	}
@@ -229,21 +249,8 @@ func (sc *gatewaySource) filterByAnnotations(gateways []*networkingv1alpha3.Gate
 	return filteredList, nil
 }
 
-func parseIngress(ingress string) (namespace, name string, err error) {
-	parts := strings.Split(ingress, "/")
-	if len(parts) == 2 {
-		namespace, name = parts[0], parts[1]
-	} else if len(parts) == 1 {
-		name = parts[0]
-	} else {
-		err = fmt.Errorf("invalid ingress name (name or namespace/name) found %q", ingress)
-	}
-
-	return
-}
-
 func (sc *gatewaySource) targetsFromIngress(ctx context.Context, ingressStr string, gateway *networkingv1alpha3.Gateway) (targets endpoint.Targets, err error) {
-	namespace, name, err := parseIngress(ingressStr)
+	namespace, name, err := utils.ParseIngress(ingressStr)
 	if err != nil {
 		return nil, fmt.Errorf("failed to parse Ingress annotation on Gateway (%s/%s): %w", gateway.Namespace, gateway.Name, err)
 	}
@@ -266,44 +273,24 @@ func (sc *gatewaySource) targetsFromIngress(ctx context.Context, ingressStr stri
 	return
 }
 
-func (sc *gatewaySource) targetsFromGateway(ctx context.Context, gateway *networkingv1alpha3.Gateway) (targets endpoint.Targets, err error) {
-	targets = getTargetsFromTargetAnnotation(gateway.Annotations)
+func (sc *gatewaySource) targetsFromGateway(ctx context.Context, gateway *networkingv1alpha3.Gateway) (endpoint.Targets, error) {
+	targets := getTargetsFromTargetAnnotation(gateway.Annotations)
 	if len(targets) > 0 {
-		return
+		return targets, nil
 	}
 
 	ingressStr, ok := gateway.Annotations[IstioGatewayIngressSource]
 	if ok && ingressStr != "" {
-		targets, err = sc.targetsFromIngress(ctx, ingressStr, gateway)
-		return
+		return sc.targetsFromIngress(ctx, ingressStr, gateway)
 	}
 
-	services, err := sc.serviceInformer.Lister().Services(sc.namespace).List(labels.Everything())
+	targets, err := utils.EndpointTargetsFromServices(sc.serviceInformer, sc.namespace, gateway.Spec.Selector)
+
 	if err != nil {
-		log.Error(err)
-		return
+		return nil, err
 	}
 
-	for _, service := range services {
-		if !gatewaySelectorMatchesServiceSelector(gateway.Spec.Selector, service.Spec.Selector) {
-			continue
-		}
-
-		if len(service.Spec.ExternalIPs) > 0 {
-			targets = append(targets, service.Spec.ExternalIPs...)
-			continue
-		}
-
-		for _, lb := range service.Status.LoadBalancer.Ingress {
-			if lb.IP != "" {
-				targets = append(targets, lb.IP)
-			} else if lb.Hostname != "" {
-				targets = append(targets, lb.Hostname)
-			}
-		}
-	}
-
-	return
+	return targets, nil
 }
 
 // endpointsFromGatewayConfig extracts the endpoints from an Istio Gateway Config object
